+++ packages/components/components/elvis-accordion/src/react/elvia-accordion.tsx
@@ -8,6 +8,8 @@ export interface AccordionProps {
   content: string | HTMLElement;
   openLabel?: string;
   closeLabel?: string;
+  openAriaLabel?: string;
+  closeAriaLabel?: string;
   labelPosition: AccordionLabelPosition;
   size: AccordionSize;
   type: AccordionType;
@@ -19,6 +21,8 @@ const Accordion: FC<AccordionProps> = ({
   content,
   openLabel,
   closeLabel,
+  openAriaLabel,
+  closeAriaLabel,
   labelPosition = 'center',
   size = 'medium',
   type = 'normal',
@@ -61,6 +65,14 @@ const Accordion: FC<AccordionProps> = ({
     }
   });
 
+  const decideButtonAriaLabel = (): string => {
+    if (contentOpen) {
+      return closeAriaLabel ? closeAriaLabel : closeLabel ? closeLabel : 'Lukk';
+    } else {
+      return openAriaLabel ? openAriaLabel : openLabel ? openLabel : 'Åpne';
+    }
+  };
+
   return (
     <span ref={accordionRef}>
       <AccordionArea
@@ -90,7 +102,7 @@ const Accordion: FC<AccordionProps> = ({
             onMouseEnter={() => setIsHoveringButton(true)}
             onMouseLeave={() => setIsHoveringButton(false)}
             data-testid="accordion-button-label"
-            aria-label={contentOpen ? 'Lukk' : 'Åpne'}
+            aria-label={decideButtonAriaLabel()}
           >
             {!contentOpen ? openLabel : closeLabel}
             <Icon